Add basic SQL Anonymizer (#15)
* Add Expression.sqlExpr to use different string representations for plan trees and SQLs

* Recursively apply expression filters

* Add TODO note

* Print anonymized SQL

+++ src/main/scala/wvlet/msgframe/sql/parser/SQLInterpreter.scala
@@ -568,7 +568,7 @@ class SQLInterpreter extends SqlBaseBaseVisitor[Any] with LogSupport {
     QuotedIdentifier(ctx.getText.replaceAll("(^\"|\"$)", ""))
   }
   override def visitDigitIdentifier(ctx: DigitIdentifierContext): Identifier = {
-    DigitId(ctx.getText.toInt)
+    DigitId(ctx.getText)
   }
 
   override def visitOver(ctx: OverContext): Window = {
